# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Lottery App API",
      "description": "Find out how your APIs work",
      "license": {
        "name": "VU",
        "url": ""
      }
    },
    "host": "localhost:8080",
    "basePath": "/api/tickets",
...
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
          "/{id}": {
            "get": {
                "summary": "Get ticket By Id",
                "description": "Get ticket By Id, returns only if ticket is active",
                "operationId": "getTicketById",
            "parameters": [
              {
                "name": "id",
                "in": "path",
                "description": "Ticket ID",
                "required": true,
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            ],
            "responses": {
              "200": {
                "description": "Successful operation",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/definitions/Ticket"
                    }
                  }
                }
              },
              "400":{
                "description": "Add lines failed!!"
              }
            },
                "produces": [
                  "application/json"
                ]
                  
                }
                },
...
          },
    "definitions": {
      "Ticket": {
        "properties": {
          "lineQuantity": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
   }
  }